+++ src/Payment/class-wsz-payment-handler.php
@@ -10,6 +10,8 @@ class WSZ_Payment_Handler
 
     private ?WSZ_PayNL_Gateway_Integration $paynl_gateway = null;
 
+    private ?WSZ_Stripe_Gateway_Integration $stripe_gateway = null;
+
     private ?WSZ_Tokenized_Gateway $tokenized_gateway = null;
 
     public function __construct(WSZ_Subscription_Manager $subscription_manager)
@@ -24,6 +26,9 @@ class WSZ_Payment_Handler
 
         $this->paynl_gateway = new WSZ_PayNL_Gateway_Integration($this->subscription_manager);
         $this->paynl_gateway->init();
+
+        $this->stripe_gateway = new WSZ_Stripe_Gateway_Integration();
+        $this->stripe_gateway->init();
 
         $this->tokenized_gateway = new WSZ_Tokenized_Gateway($this->subscription_manager, $this);
         $this->tokenized_gateway->init();